requirements.txt
Pillow
tqdm
rich